# Clean up and optimise MemoryExtensions