--- a/src/DSE.Open/MemoryExtensions.Byte.cs
+++ b/src/DSE.Open/MemoryExtensions.Byte.cs
@@ -2,11 +2,22 @@
 // Down Syndrome Education International and Contributors licence this file to you under the MIT license.
 
 using System.Buffers;
+using System.Text;
 
 namespace DSE.Open;
 
 public static partial class MemoryExtensions
 {
+    internal static class SearchBytes
+    {
+        internal static readonly SearchValues<byte> s_asciiLetters = SearchValues.Create("ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz"u8);
+
+        internal static readonly SearchValues<byte> s_asciiLettersAndDigits =
+            SearchValues.Create("ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789"u8);
+
+        internal static readonly SearchValues<byte> s_asciiUpperLettersAndDigits = SearchValues.Create("ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"u8);
+    }
+
     public static bool ContainsOnlyAsciiDigits(this Span<byte> value)
         => ContainsOnlyAsciiDigits((ReadOnlySpan<byte>)value);
 
@@ -19,17 +30,30 @@ public static partial class MemoryExtensions
     public static bool ContainsOnlyAsciiUpperLettersOrDigits(this Span<byte> value)
         => ContainsOnlyAsciiUpperLettersOrDigits((ReadOnlySpan<byte>)value);
 
+    public static bool ContainsOnlyAsciiLettersLower(this Span<byte> value)
+        => ContainsOnlyAsciiLettersLower((ReadOnlySpan<byte>)value);
+
+    public static bool ContainsOnlyAsciiLettersUpper(this Span<byte> value)
+        => ContainsOnlyAsciiLettersUpper((ReadOnlySpan<byte>)value);
+
     public static bool ContainsOnlyAsciiDigits(this ReadOnlySpan<byte> value)
-        => ContainsOnly(value, AsciiChar.IsDigit);
+        => value.IndexOfAnyExceptInRange((byte)'0', (byte)'9') == -1;
 
     public static bool ContainsOnlyAsciiLetters(this ReadOnlySpan<byte> value)
-        => ContainsOnly(value, AsciiChar.IsLetter);
+        => value.IndexOfAnyExcept(SearchBytes.s_asciiLetters) == -1;
 
     public static bool ContainsOnlyAsciiLettersOrDigits(this ReadOnlySpan<byte> value)
-        => ContainsOnly(value, AsciiChar.IsLetterOrDigit);
+        => value.IndexOfAnyExcept(SearchBytes.s_asciiLettersAndDigits) == -1;
 
     public static bool ContainsOnlyAsciiUpperLettersOrDigits(this ReadOnlySpan<byte> value)
-        => ContainsOnly(value, v => AsciiChar.IsLetterUpper(v) || AsciiChar.IsDigit(v));
+        => value.IndexOfAnyExcept(SearchBytes.s_asciiUpperLettersAndDigits) == -1;
+
+    public static bool ContainsOnlyAsciiLettersLower(this ReadOnlySpan<byte> value)
+        => value.IndexOfAnyExceptInRange((byte)'a', (byte)'z') == -1;
+
+    public static bool ContainsOnlyAsciiLettersUpper(this ReadOnlySpan<byte> value)
+        => value.IndexOfAnyExceptInRange((byte)'A', (byte)'Z') == -1;
+
 
     public static bool TryCopyWhereNotWhitespace(this Span<byte> span, Span<byte> buffer, out int bytesWritten)
         => TryCopyWhereNotWhitespace((ReadOnlySpan<byte>)span, buffer, out bytesWritten);
@@ -73,6 +97,7 @@ public static partial class MemoryExtensions
     public static Span<byte> RemoveNonAsciiLetterOrDigit(this Span<byte> span)
         => Remove(span, v => !AsciiChar.IsLetterOrDigit(v));
 
+    [Obsolete("Use `Ascii.ToLowerInPlace`")]
     public static void ToLower(this Span<byte> span)
     {
         if (span.IsEmpty)
@@ -86,6 +111,7 @@ public static partial class MemoryExtensions
         }
     }
 
+    [Obsolete("Use `Ascii.ToUpperInPlace`")]
     public static void ToUpper(this Span<byte> span)
     {
         if (span.IsEmpty)
@@ -99,143 +125,28 @@ public static partial class MemoryExtensions
         }
     }
 
-    public static bool ContainsOnlyLetters(ReadOnlySpan<AsciiChar> value)
-    {
-        for (var i = 0; i < value.Length; i++)
-        {
-            if (!AsciiChar.IsLetter(value[i]))
-            {
-                return false;
-            }
-        }
+    /// <summary>
+    /// Checks if the span contains only ASCII letters.
+    /// </summary>
+    /// <param name="value"></param>
+    /// <returns></returns>
+    public static bool ContainsOnlyLetters(this ReadOnlySpan<AsciiChar> value)
+        => ValuesMarshal.AsBytes(value).ContainsOnlyAsciiLetters();
 
-        return true;
-    }
+    /// <summary>
+    /// Checks if the span contains only valid ASCII bytes.
+    /// </summary>
+    /// <param name="value"></param>
+    /// <returns></returns>
+    public static bool IsAscii(this ReadOnlySpan<byte> value) => Ascii.IsValid(value);
 
-    public static bool ContainsOnlyLetters(ReadOnlySpan<byte> value)
-    {
-        for (var i = 0; i < value.Length; i++)
-        {
-            if (!AsciiChar.IsLetter(value[i]))
-            {
-                return false;
-            }
-        }
 
-        return true;
-    }
-
-    public static bool IsAscii(ReadOnlySpan<byte> value)
-    {
-        for (var i = 0; i < value.Length; i++)
-        {
-            if (!AsciiChar.IsAscii(value[i]))
-            {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    public static bool SequenceEqualsCaseInsensitive(ReadOnlySpan<byte> a, ReadOnlySpan<byte> b)
-    {
-        if (a.Length != b.Length)
-        {
-            return false;
-        }
-
-        for (var i = 0; i < a.Length; i++)
-        {
-            if (!AsciiChar.EqualsCaseInsensitive(a[i], b[i]))
-            {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    public static bool SequenceEqualsCaseInsensitive(ReadOnlySpan<AsciiChar> a, ReadOnlySpan<AsciiChar> b)
-    {
-        if (a.Length != b.Length)
-        {
-            return false;
-        }
-
-        for (var i = 0; i < a.Length; i++)
-        {
-            if (!AsciiChar.EqualsCaseInsensitive(a[i], b[i]))
-            {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    public static void ConvertToChar(ReadOnlySpan<byte> source, Span<char> output, out int charsWritten)
-    {
-        if (source.Length > output.Length)
-        {
-            ThrowHelper.ThrowInvalidOperationException("Output buffer must be at least as large as source.");
-            charsWritten = 0; // unreachable
-            return;
-        }
-
-        for (var i = 0; i < source.Length; i++)
-        {
-            output[i] = (char)source[i];
-        }
-
-        charsWritten = source.Length;
-    }
-
-    public static char[] ToCharArray(ReadOnlySpan<byte> source)
-    {
-        char[]? rented = null;
-
-        try
-        {
-            Span<char> chars = source.Length <= StackallocThresholds.MaxCharLength
-                ? stackalloc char[StackallocThresholds.MaxCharLength]
-                : (rented = ArrayPool<char>.Shared.Rent(source.Length));
-
-            ConvertToChar(source, chars, out var charsWritten);
-            return chars[..charsWritten].ToArray();
-        }
-        finally
-        {
-            if (rented is not null)
-            {
-                ArrayPool<char>.Shared.Return(rented);
-            }
-        }
-    }
-
-    public static ReadOnlySpan<byte> ToByteSpan(ReadOnlySpan<char> source)
-    {
-        byte[]? rented = null;
-
-        try
-        {
-            Span<byte> chars = source.Length <= StackallocThresholds.MaxByteLength
-                ? stackalloc byte[StackallocThresholds.MaxByteLength]
-                : (rented = ArrayPool<byte>.Shared.Rent(source.Length));
-
-            for (var i = 0; i < source.Length; i++)
-            {
-                chars[i] = (byte)source[i];
-            }
-
-            return chars[..source.Length].ToArray();
-        }
-        finally
-        {
-            if (rented is not null)
-            {
-                ArrayPool<byte>.Shared.Return(rented);
-            }
-        }
-    }zx
+    /// <summary>
+    /// Determines whether the specified spans of ASCII are equal, ignoring case.
+    /// </summary>
+    /// <param name="a"></param>
+    /// <param name="b"></param>
+    /// <returns></returns>
+    public static bool SequenceEqualsCaseInsensitive(this ReadOnlySpan<AsciiChar> a, ReadOnlySpan<AsciiChar> b)
+        => Ascii.EqualsIgnoreCase(ValuesMarshal.AsBytes(a), ValuesMarshal.AsBytes(b));
 }