… benchmarks/DSE.Open.Benchmarks/MemoryExtensions/Chars/ContainsOnlyAsciiDigitsCharBenchmarks.cs
… benchmarks/DSE.Open.Benchmarks/MemoryExtensions/Chars/ContainsOnlyAsciiLettersCharBenchmarks.cs
… src/DSE.Open/AsciiString.Helpers.cs