## Using `IndexOfAnyExceptInRange`

The pattern here is that `IndexOfAnyExceptInRange` is marginally slower than a simple loop when the first opportunity to fail is early in the span. However, by the time the first match is 8th in the span the vectorized version is significantly faster.

## Bytes

| Method                           | arr       |       Mean |
| -------------------------------- | --------- | ---------: |
| ContainsOnlyAsciiDigits_Original | Byte[1]   |  0.2413 ns |
| ContainsOnlyAsciiDigits          | Byte[1]   |  1.7887 ns |
|                                  |           |            |
| ContainsOnlyAsciiDigits_Original | Byte[8]   |  3.9190 ns |
| ContainsOnlyAsciiDigits          | Byte[8]   |  5.8323 ns |
|                                  |           |            |
| ContainsOnlyAsciiDigits_Original | Byte[109] | 77.7796 ns |
| ContainsOnlyAsciiDigits          | Byte[109] |  4.5177 ns |

### Chars

| Method                           | str                 |       Mean |     Error |    StdDev | Ratio | RatioSD |
| -------------------------------- | ------------------- | ---------: | --------: | --------: | ----: | ------: |
| ContainsOnlyAsciiDigits_Original | a                   |  0.7044 ns | 0.0076 ns | 0.0067 ns |  1.00 |    0.00 |
| ContainsOnlyAsciiDigits          | a                   |  2.1530 ns | 0.0166 ns | 0.0138 ns |  3.06 |    0.04 |
|                                  |                     |            |           |           |       |         |
| ContainsOnlyAsciiDigits_Original | 1234567a            |  4.0428 ns | 0.1052 ns | 0.1169 ns |  1.00 |    0.00 |
| ContainsOnlyAsciiDigits          | 1234567a            |  3.9195 ns | 0.0236 ns | 0.0209 ns |  0.97 |    0.03 |
|                                  |                     |            |           |           |       |         |
| ContainsOnlyAsciiDigits_Original | 1234(...)789a [109] | 79.8199 ns | 1.2072 ns | 1.1292 ns |  1.00 |    0.00 |
| ContainsOnlyAsciiDigits          | 1234(...)789a [109] |  6.0401 ns | 0.0435 ns | 0.0385 ns |  0.08 |    0.00 |

```cs
[MemoryDiagnoser]
public class ContainsOnlyAsciiDigitsBenchmarks
{
    private static readonly byte[] Easy = "a"u8.ToArray();

    private static readonly byte[] Medium = "1234567a"u8.ToArray();

    private static readonly byte[] Hard = "1234567891   23456789123456789123456789123456789123456789123456789123456789123456789123456789123456789123456789a"u8.ToArray();

    public static IEnumerable<object> Params()
    {
        yield return Easy;
        yield return Medium;
        yield return Hard;
    }

    [Benchmark(Baseline = true)]
    [ArgumentsSource(nameof(Params))]
    public bool ContainsOnlyAsciiDigits_Original(byte[] arr)
    {
        var value = arr.AsSpan();

        for (var i = 0; i < value.Length; i++)
        {
            if (!AsciiChar.IsDigit(value[i]))
            {
                return false;
            }
        }

        return true;
    }

    [Benchmark]
    [ArgumentsSource(nameof(Params))]
    public bool ContainsOnlyAsciiDigits(byte[] arr) => arr.AsSpan().ContainsOnlyAsciiDigits();
}
```

## Using `SearchValues`

`SearchValues` also slightly worse when the first opportunity to fail is very early, but again, pulls away at any reasonable size.

### Bytes

| Method                            | arr       |       Mean |
| --------------------------------- | --------- | ---------: |
| ContainsOnlyAsciiLetters_Original | Byte[1]   |  0.3718 ns |
| ContainsOnlyAsciiLetters          | Byte[1]   |  2.0296 ns |
|                                   |           |            |
| ContainsOnlyAsciiLetters_Original | Byte[8]   |  4.3940 ns |
| ContainsOnlyAsciiLetters          | Byte[8]   |  2.7138 ns |
|                                   |           |            |
| ContainsOnlyAsciiLetters_Original | Byte[105] | 77.1056 ns |
| ContainsOnlyAsciiLetters          | Byte[105] |  5.2760 ns |

```cs
[MemoryDiagnoser]
public class ContainsOnlyAsciiLettersBenchmarks
{
    private static readonly byte[] s_easy = "a"u8.ToArray();

    private static readonly byte[] s_medium = "abcdEFG1"u8.ToArray();

    private static readonly byte[] s_hard = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1"u8
        .ToArray();

    public static IEnumerable<object> Params()
    {
        yield return s_easy;
        yield return s_medium;
        yield return s_hard;
    }

    [Benchmark(Baseline = true)]
    [ArgumentsSource(nameof(Params))]
    public bool ContainsOnlyAsciiLetters_Original(byte[] arr)
    {
        var value = arr.AsSpan();

        for (var i = 0; i < value.Length; i++)
        {
            if (!AsciiChar.IsLetter(value[i]))
            {
                return false;
            }
        }

        return true;
    }

    [Benchmark]
    [ArgumentsSource(nameof(Params))]
    public bool ContainsOnlyAsciiLetters(byte[] arr) => arr.AsSpan().ContainsOnlyAsciiLetters();
}
```

### Chars

| Method                            | str                 |       Mean |     Error |    StdDev | Ratio | RatioSD |
| --------------------------------- | ------------------- | ---------: | --------: | --------: | ----: | ------: |
| ContainsOnlyAsciiLetters_Original | a                   |  0.4964 ns | 0.0305 ns | 0.0255 ns |  1.00 |    0.00 |
| ContainsOnlyAsciiLetters          | a                   |  2.3055 ns | 0.0212 ns | 0.0177 ns |  4.66 |    0.25 |
|                                   |                     |            |           |           |       |         |
| ContainsOnlyAsciiLetters_Original | abcdEFG1            |  4.8851 ns | 0.0759 ns | 0.0672 ns |  1.00 |    0.00 |
| ContainsOnlyAsciiLetters          | abcdEFG1            |  3.0579 ns | 0.0428 ns | 0.0401 ns |  0.63 |    0.01 |
|                                   |                     |            |           |           |       |         |
| ContainsOnlyAsciiLetters_Original | abcd(...)XYZ1 [105] | 80.7324 ns | 1.5950 ns | 1.4920 ns |  1.00 |    0.00 |
| ContainsOnlyAsciiLetters          | abcd(...)XYZ1 [105] |  5.7671 ns | 0.0558 ns | 0.0494 ns |  0.07 |    0.00 |