@rabuckley 2023-09-14
## Bytes

| Method                           | arr       |      Mean |
| -------------------------------- | --------- | --------: |
| ContainsOnlyAsciiDigits_Original | Byte[1]   | 0.1564 ns |
| ContainsOnlyAsciiDigits          | Byte[1]   | 0.7968 ns |
|                                  |           |           |
| ContainsOnlyAsciiDigits_Original | Byte[8]   | 2.9687 ns |
| ContainsOnlyAsciiDigits          | Byte[8]   | 3.6202 ns |
|                                  |           |           |
| ContainsOnlyAsciiDigits_Original | Byte[112] | 3.9344 ns |
| ContainsOnlyAsciiDigits          | Byte[112] | 2.5195 ns |

| Method                            | arr       |       Mean |
| --------------------------------- | --------- | ---------: |
| ContainsOnlyAsciiLetters_Original | Byte[1]   |  0.1759 ns |
| ContainsOnlyAsciiLetters          | Byte[1]   |  0.8838 ns |
|                                   |           |            |
| ContainsOnlyAsciiLetters_Original | Byte[8]   |  8.2088 ns |
| ContainsOnlyAsciiLetters          | Byte[8]   |  3.9375 ns |
|                                   |           |            |
| ContainsOnlyAsciiLetters_Original | Byte[105] | 49.3615 ns |
| ContainsOnlyAsciiLetters          | Byte[105] |  6.5345 ns |

## Chars

| Method                           | str                 |       Mean |     Error |    StdDev | Ratio |
| -------------------------------- | ------------------- | ---------: | --------: | --------: | ----: |
| ContainsOnlyAsciiDigits_Original | 1234(...)789a [109] | 42.8465 ns | 0.2105 ns | 0.1866 ns |  1.00 |
| ContainsOnlyAsciiDigits          | 1234(...)789a [109] |  7.9192 ns | 0.0501 ns | 0.0468 ns |  0.18 |
|                                  |                     |            |           |           |       |
| ContainsOnlyAsciiDigits_Original | 1234567a            |  8.4434 ns | 0.1661 ns | 0.1554 ns |  1.00 |
| ContainsOnlyAsciiDigits          | 1234567a            |  2.3705 ns | 0.0065 ns | 0.0054 ns |  0.28 |
|                                  |                     |            |           |           |       |
| ContainsOnlyAsciiDigits_Original | a                   |  0.5050 ns | 0.0025 ns | 0.0024 ns |  1.00 |
| ContainsOnlyAsciiDigits          | a                   |  0.8165 ns | 0.0035 ns | 0.0032 ns |  1.62 |

| Method                            | str                 |       Mean |     Error |    StdDev | Ratio | RatioSD |
| --------------------------------- | ------------------- | ---------: | --------: | --------: | ----: | ------: |
| ContainsOnlyAsciiLetters_Original | a                   |  0.1972 ns | 0.0089 ns | 0.0079 ns |  1.00 |    0.00 |
| ContainsOnlyAsciiLetters          | a                   |  1.1263 ns | 0.0059 ns | 0.0055 ns |  5.72 |    0.23 |
|                                   |                     |            |           |           |       |         |
| ContainsOnlyAsciiLetters_Original | abcdEFG1            |  8.3313 ns | 0.0245 ns | 0.0229 ns |  1.00 |    0.00 |
| ContainsOnlyAsciiLetters          | abcdEFG1            |  2.6875 ns | 0.0107 ns | 0.0095 ns |  0.32 |    0.00 |
|                                   |                     |            |           |           |       |         |
| ContainsOnlyAsciiLetters_Original | abcd(...)XYZ1 [105] | 47.7211 ns | 0.3031 ns | 0.2531 ns |  1.00 |    0.00 |
| ContainsOnlyAsciiLetters          | abcd(...)XYZ1 [105] |  7.2375 ns | 0.0203 ns | 0.0190 ns |  0.15 |    0.00 |